tagpr: separation of concerns
move service parts to service, and util to util file
